Move formatIsoString() to utils
This might be used by multiple parts of the project for formatting
dates. Move it to utils.

This also changes the month to use the 3-letter abbreviation.
